backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
fastapi==0.132.0
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==26.0
postgrest==2.28.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.28.0
requests==2.32.5
rich==14.3.3
six==1.17.0
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0